Add files via upload
Still needs a lot of work. Going to change the note style similar to the one in Zack's get_data_file.  Going to create functions that takes the csv file information from Zack's script. Going to convert it into a class file so that the graphing functions can be called from the get_data_file.py. Going to create a line graph option of the temperature from year to year to see if there's correlation from bleaching and temperature.